Box was not being configured properly.
Env files were being copied to a non-existent directory, so the copy command was failing and the bridge and dapp were not being configured. Updated the destination for these.

The setup shell script will now fail if any command fails. This should prevent this kind of error in the future.

Git commands have been tightened.